some code replaced with better form. logicals to impute, removing if statements. 2 unnecessary loops removed as well.
Noticed potential bug, still need to fix:
if there are 0s at the end of Mx (e.g. in early years included in HMD),
it would make sense so have lx equal zero by that time, i.e. put on the
kill qx of 1 prior to N. This could be annoying by implying a large
amount of terminal imputing, BUT appears preferable to insanely high
remaining life expectancies in the oldedst ages, which is what can
happen now.